repackage new robot code to work with robot multiplexing infrastructure
