--- robots/preservation/item_error.rb
@@ -1,7 +1,0 @@
-module Preservation
-
-  # ItemError wraps a causal exception, creating a new exception
-  #   that usually terminates processing of the current item
-  class ItemError < StandardError
-  end
-end
