+++ spec/preservation_ingest/base_spec.rb
@@ -1,4 +1,4 @@
-describe Preservation::Base do
+describe Robots::SdrRepo::PreservationIngest::Base do
   describe '.execute_shell_command' do
     it 'returns stdout for success' do
       cmd = 'echo "hello world!"'
